requirements.txt
pynput>=1.7.6
pyaudio>=0.2.11
openai-whisper>=20231117
librosa>=0.10.0
soundfile>=0.12.1
faster-whisper>=1.1.1
pyautogui>=0.9.54
Pillow>=9.5.0
llama-cpp-python>=0.2.90
huggingface-hub>=0.23.0
websockets>=11.0.3